% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_waterbody_area.R
\name{add_waterbody_area}
\alias{add_waterbody_area}
\title{Title Add a column for area to a table with waterbody (lake or river) names}
\usage{
add_waterbody_area(
  dat,
  waterbody_polygons,
  units = c("sq_meters", "sq_km", "acres"),
  digits = 2
)
}
\arguments{
\item{dat}{A table that includes a column of waterbody names}

\item{waterbody_polygons}{sf table of waterbody polygons (perhaps filtered for unique names of your dat!)}

\item{units}{Units for area measurement; one of 'sq_meters', 'sq_km', or 'acres'}

\item{digits}{How many digits to include in area measurement}
}
\value{
Your table with a new column 'wb_area'
}
\description{
Title Add a column for area to a table with waterbody (lake or river) names
}
