% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_coordinates.R
\name{clean_coordinates}
\alias{clean_coordinates}
\title{Title Clean latitude and longitude decimal degree coordinate data.}
\usage{
clean_coordinates(dat)
}
\arguments{
\item{dat}{A dataset with columns containing latitude and longitude in decimal degree format.}
}
\value{
A dataset with common typos in latitude / longitude columns corrected.
}
\description{
Title Clean latitude and longitude decimal degree coordinate data.
}
\examples{
dat = data.frame(Latitude = c('48.19N','38.92','-112.10'), Longitude = c('110.392W','-123.11','49.32'))
clean_coordinates(dat)

}
