% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_waterbody_names.R
\name{clean_waterbody_names}
\alias{clean_waterbody_names}
\title{Title Clean the waterbody name column of a table}
\usage{
clean_waterbody_names(dat, waterbody_column)
}
\arguments{
\item{dat}{A dataset with waterbody names that need cleaning.}

\item{waterbody_column}{The column name (in quotations) that encodes the waterbody names.}
}
\value{
A dataset with common typos / different naming conventions in waterbody names corrected.
}
\description{
Title Clean the waterbody name column of a table
}
\examples{
dat = data.frame(watname = c("Lac La Hache","Shuswap Lake","Koocanusa Lake"))
clean_waterbody_names(dat)
}
