% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_risk_layers_available.R
\name{current_risk_layers_available}
\alias{current_risk_layers_available}
\title{Title Get current list of risk layers hard-coded into prioritization model}
\usage{
current_risk_layers_available()
}
\value{
A vector of risk layer names
}
\description{
Title Get current list of risk layers hard-coded into prioritization model
}
\examples{
current_risk_layers_available()
}
