% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_species_in_waterbody.R
\name{find_all_species_in_waterbody}
\alias{find_all_species_in_waterbody}
\title{Title Find all species occurrence records in a waterbody}
\usage{
find_all_species_in_waterbody(
  wb,
  in_shiny = F,
  sources = c("SPI", "Old Aquatic", "Incident Reports", "iNaturalist"),
  exclude = c("Fungi", "Plantae"),
  excel_path = "5_Incidental Observations/Master Incidence Report Records.xlsx",
  sheet_name = "Aquatic Reports",
  excel_species_var = "Submitted_Common_Name",
  output_crs = 4326,
  quiet = T,
  ...
)
}
\arguments{
\item{wb}{Waterbody in which to search}

\item{in_shiny}{Is this being run in a 'shiny' environment?}

\item{sources}{Which data sources should we look in? One or more of 'SPI','Old Aquatic','Incident Reports', and 'iNaturalist'}

\item{exclude}{Which kinds of organisms should we exclude? Default is 'Fungi' and 'Plantae'}

\item{excel_path}{Path (including excel file name) to master incident tracking sheet}

\item{sheet_name}{Name of aquatic sheet in excel file}

\item{excel_species_var}{Name of column with species common name in excel sheet}

\item{output_crs}{Output coordinate reference system (e.g. WGS 84, i.e. lat/long)}

\item{quiet}{Should this function run in a verbose fashion?}

\item{...}{}
}
\value{
A table of species occurrence data
}
\description{
Title Find all species occurrence records in a waterbody
}
\examples{
\dontrun{}
}
