% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_downstream_course.R
\name{find_downstream_course}
\alias{find_downstream_course}
\title{Use the BC Freshwater Atlas to trace course downstream from chosen waterbody polygon}
\usage{
find_downstream_course(
  wb_polygon,
  wb_is_lake = T,
  wb_separate = T,
  ggplot_fig = F
)
}
\arguments{
\item{wb_polygon}{Chosen waterbody polygon;}

\item{wb_is_lake}{Is the chosen waterbody polygon a lake?}

\item{wb_separate}{Would you like the resulting series of linestrings to be joined into one multilinestring?}

\item{ggplot_fig}{Should a ggplot figure be made an appended to output list?}
}
\value{
An sf vector table of downstream waterways; optional ggplot figure.
}
\description{
Use the BC Freshwater Atlas to trace course downstream from chosen waterbody polygon
}
\examples{
\dontrun {
shus = bcdata::bcdc_query_geodata('freshwater-atlas-lakes') |>
bcdata::filter(GNIS_NAME_1 == 'Shuswap Lake') |>
bcdata::collect()

find_downstream_course(shus, ggplot_fig = T)
}

}
