% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bcdc_risk_layer.R
\name{get_bcdc_risk_layer}
\alias{get_bcdc_risk_layer}
\title{Title Download / calculate risk layers from BC Data Catalogue}
\usage{
get_bcdc_risk_layer(geog_units, geog_id_col, layer = c("prox_to_settlements"))
}
\arguments{
\item{geog_units}{Geographic units for which to calculate risk layers.}

\item{layer}{Risk layer; one of 'prox_to_settlements'}
}
\value{
Various datasets that can be plugged into R6 priorization models as risk layers.
}
\description{
Title Download / calculate risk layers from BC Data Catalogue
}
\examples{
\dontrun{}
}
