% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected_waterbodies.R
\name{get_connected_waterbodies}
\alias{get_connected_waterbodies}
\title{Title}
\usage{
get_connected_waterbodies(
  waterbody_name = NULL,
  waterbody_coordinates = NULL,
  waterbody_polygon = NULL,
  waterbody_type = "lake",
  search_radius = 10,
  in_shiny = FALSE,
  the_session = session,
  quiet = T
)
}
\arguments{
\item{waterbody_name}{Name of the waterbody on which to base a connected network of waterbodies}

\item{waterbody_coordinates}{(Required unless polygon included) vector of WGS 84 coordinates, in format: c(longitude, latitude)}

\item{waterbody_polygon}{(Optional) Polygon of your chosen waterbody; include if you have one!}

\item{waterbody_type}{For now, only lakes are implemented}

\item{search_radius}{How far from the waterbody to search (in kilometers)}

\item{in_shiny}{Is this being run in a 'shiny' environment?}

\item{the_session}{What is the shiny session?}

\item{quiet}{Should this function return copious updates as it runs?}
}
\value{
A spatial table (sf object) of all lakes, streams and rivers that intersect with your chosen waterbody
}
\description{
Title
}
\examples{
get_connected_waterbodies('Stave Lake',
waterbody_coordinates = c(-122.2837065448555,49.37115621966073),
quiet = F)
}
