% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_natural_resource_area_shapefile.R
\name{get_natural_resource_area_shapefile}
\alias{get_natural_resource_area_shapefile}
\title{Title Download the polygon(s) for a natural resource region or district, by name}
\usage{
get_natural_resource_area_shapefile(
  area_type = area_type,
  specific_areas = "all"
)
}
\arguments{
\item{area_type}{The natural resource are type; one of 'region' or 'district'}

\item{specific_areas}{Either 'all' or a concatenated vector of region or district names as strings.}
}
\value{
A spatial table (sf object) of the region(s) or district(s) specified.
}
\description{
Title Download the polygon(s) for a natural resource region or district, by name
}
\examples{
if(interactive()){
get_natural_resource_area_shapefile('regions', c('Thompson-Okanagan','South Coast'))
}
}
