% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterbodies_in_area.R
\name{get_waterbodies_in_area}
\alias{get_waterbodies_in_area}
\title{Title}
\usage{
get_waterbodies_in_area(
  area_type = c("regions", "districts"),
  specific_areas = c(NULL),
  size_threshold = 100,
  output_crs = 4326
)
}
\arguments{
\item{area_type}{Natural Resource area type; one of 'regions' or 'districts'}

\item{specific_areas}{The name(s) of one or more polygons of interest from the Natural Resource area selected}

\item{size_threshold}{Minimum area in square kilometers of lakes to include in download}

\item{output_crs}{Spatial projection system to use for lake sf object; defaults to WGS 84 (EPSG:4326)}
}
\value{
An sf object (spatial table) of lakes in the selected polygon(s) within the chosen Natural Resource area
}
\description{
Title
}
\examples{
get_waterbodies_in_area(area_type = 'regions',
specific_areas = c("Thompson-Okanagan","West Coast"),
size_threshold = 100000)
}
