% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterbody_polygon.R
\name{get_waterbody_polygon}
\alias{get_waterbody_polygon}
\title{Title Get the polygon for a waterbody}
\usage{
get_waterbody_polygon(focus_wb_name = NULL, focus_wb_coordinates = NULL)
}
\arguments{
\item{focus_wb_name}{Name of waterbody}

\item{focus_wb_coordinates}{Coordinates (long, lat) of waterbody}
}
\value{
A single-rowed spatial table (sf object)
}
\description{
Title Get the polygon for a waterbody
}
\examples{
/dontrun{}
}
