% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_aq_occ_data.R
\name{grab_aq_occ_data}
\alias{grab_aq_occ_data}
\title{Grab Aquatic BC Occurrence Data}
\usage{
grab_aq_occ_data(
  common_names = NULL,
  excel_path =
    "J:/2 SCIENCE - Invasives/SPECIES/5_Incidental Observations/Master Incidence Report Records.xlsx",
  sheet_name = "Aquatic Reports",
  excel_species_var = NULL,
  output_crs = 4326,
  quiet = T,
  ...
)
}
\arguments{
\item{common_names}{A vector of common names for one or more species of interest.}

\item{excel_path}{Optional; path to your excel file (must include columns Date, Species, Scientific, Location, Latitude and Longitude)}

\item{sheet_name}{Optional; if you read in your own excel file, what is the excel sheet name?}

\item{excel_species_var}{Optional; if you read in your own excel file, what is the name of the column listing common names?}

\item{output_crs}{Coordinate Reference System (i.e. projection system); defaults to 4326 (WGS 84), another common option for BC is 3005.}

\item{quiet}{Boolean to determine amount of feedback given by function}

\item{...}{Additional arguments}
}
\value{
Aquatic occurrence data in British Columbia; optional to add in one's own excel file from local machine.
}
\description{
Grab Aquatic BC Occurrence Data
}
\examples{
# Search for a single species (perhaps with various possible spellings)
pike_occ = grab_aq_occ_data(common_names = c("pike","northern pike"))

# Search for multiple species in one go!
invasive_fish = grab_aq_occ_data(common_names = c("black crappie","bullhead",
"black bullhead","brown bullhead","yellow bullhead"))
}
