% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_terr_occ_data.R
\name{grab_terr_occ_data}
\alias{grab_terr_occ_data}
\title{Grab Terrestrial BC Occurrence Data}
\usage{
grab_terr_occ_data(
  common_name = NULL,
  scientific_name = NULL,
  excel_path =
    "J:/2 SCIENCE - Invasives/SPECIES/5_Incidental Observations/Master Incidence Report Records.xlsx",
  sheet_name = "Aquatic Reports",
  excel_species_var = "Species",
  output_crs = 4326,
  quiet = F,
  ...
)
}
\arguments{
\item{common_name}{A vector of common names for one or more species of interest.}

\item{scientific_name}{A vector of scientific names for one or more species of interest.}

\item{excel_path}{Optional; path to your excel file (must include columns Date, Species, Scientific, Location, Latitude and Longitude)}

\item{sheet_name}{Optional; if you read in your own excel file, what is the excel sheet name?}

\item{excel_species_var}{Optional; if you read in your own excel file, what is the name of the column listing common names?}

\item{output_crs}{Coordinate Reference System (i.e. projection system); defaults to 4326 (WGS 84), another common option for BC is 3005.}

\item{quiet}{Boolean to determine amount of feedback given by function}

\item{...}{Additional arguments}
}
\value{
Terrestrial occurrence data in British Columbia; optional to add in one's own excel file from local machine.
}
\description{
Grab Terrestrial BC Occurrence Data
}
\examples{
# Search for a single species (perhaps with various possible spellings)
liz_occ = grab_terr_occ_data(common_names = c("wall lizard","common wall lizard"),
scientific_name = 'Podarcis muralis')

# Search for multiple species in one go! Here we look for wall lizards and Eastern grey squirrels.
invasives = grab_terr_occ_data(common_names = c("common wall lizard"),
scientific_name = c('Sciurus carolinensis'))

}
