% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intuit_lat_col.R
\name{intuit_lat_col}
\alias{intuit_lat_col}
\title{Title Intuit the likely column containing latitude}
\usage{
intuit_lat_col(dat)
}
\arguments{
\item{dat}{tabular data with (one) column of latitude data}
}
\value{
Returns the name of the (first) column likely containing latitude
}
\description{
Title Intuit the likely column containing latitude
}
\examples{
dat = data.frame(LaTiTuDe = c(49.12,31.22,-120.8))
intuit_lat_col(dat)
}
