% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intuit_lng_col.R
\name{intuit_lng_col}
\alias{intuit_lng_col}
\title{Title Intuit the likely column containing longitude}
\usage{
intuit_lng_col(dat, hemisphere = "W")
}
\arguments{
\item{dat}{tabular data with (one) column of longitude data}
}
\value{
Returns the name of the (first) column likely containing longitude
}
\description{
Title Intuit the likely column containing longitude
}
\examples{
dat = data.frame(lat = c(-29.112,-40.32,-41.32),
longitude = c(-120.8, -110.21, -90))
intuit_lng_col(dat)
}
