% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization_model_utils.R
\name{intuit_species_name}
\alias{intuit_species_name}
\title{Title}
\usage{
intuit_species_name(species_name, species_occurrence_data, quiet)
}
\arguments{
\item{species_name}{Name of species for which the model is run}

\item{species_occurrence_data}{Species occurrence dataset for species of interest}
}
\description{
Title
}
