% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizards}
\alias{lizards}
\title{Lizards toy dataset}
\format{
A simple feature sensu {sf}, that is to say, a spatial tidy table.
\describe
\item{DataSource}{Source layer in BCG Warehouse or Excel file}
\item{Date}{Occurrence date}
\item{Species}{Common name}
\item{Scientific}{Scientific name}
\item{Location}{Text description of occurrence location}
\item{geometry}{simple feature point geometry}
}
\source{
{Created in-house using grab_terr_occ_data function}
}
\usage{
lizards
}
\description{
20 rows of lizard occurrence data grabbed from the BCG Warehouse.
}
\keyword{datasets}
