% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occdat_leafmap.R
\name{occdat_leafmap}
\alias{occdat_leafmap}
\title{Title}
\usage{
occdat_leafmap(
  dat,
  species_column = "Species",
  location_variable = NULL,
  popup_or_label = "label",
  choose_baselayer = NULL,
  species_circles_palette = "Dark2",
  bg_regions = NULL,
  label_bg_regions = F,
  bg_region_palette = "RdBu",
  set_leaf_view = NULL,
  take_snapshot = FALSE,
  folder_for_snapshot = NULL,
  snapshot_name = NULL,
  ...
)
}
\arguments{
\item{dat}{Occurrence data that is either an sf object or has columns labelled Latitude and Longitude.}

\item{species_column}{Name of column that lists the Species' name.}

\item{location_variable}{Name of column that lists a short description of the occurrence location.}

\item{popup_or_label}{Choice of either popups or labels for leaflet map - defaults to labels.}

\item{choose_baselayer}{Optional; Select one of 'CartoDB', 'Streets' or 'Terrain.'}

\item{species_circles_palette}{Colour palette for occurrence points (e.g. 'Dark2', 'Spectral', etc.) - defaults to 'Dark2'.}

\item{bg_regions}{Optional; choose either natural resource polygons, either 'regions' or 'districts'.}

\item{label_bg_regions}{Optional; Add names to background regions, either T or F.}

\item{bg_region_palette}{Optional; Colour palette of background regions (e.g. 'Dark2', 'Spectral', etc.)}

\item{set_leaf_view}{Optional; Set custom view for map - requires concatenated vector of latitude, longitude, and zoom factor.}

\item{take_snapshot}{Optional; take snapshot of leaflet map and save to local machine, either T or F.}

\item{folder_for_snapshot}{Optional; set folder for snapshot of leaflet map.}

\item{snapshot_name}{Optional; Set name for snapshot .PNG file.}

\item{...}{Additional arguments}
}
\value{
A leaflet map of BC showing species occurrence points, optionally adding background regions and taking a snapshot .PNG file.
}
\description{
Title
}
\examples{
# using the toy dataset 'lizards', make a map!

my_map = occdat_leafmap(dat = lizards, location_variable = 'Location', bg_regions = 'regions')
my_map

# Take a snapshot of the map and save it to disk.

map_with_png = occdat_leafmap(dat = goldfish, location_variable = 'Location',
bg_regions = 'districts', bg_region_palette = 'Spectral',
take_snapshot = TRUE, folder_for_snapshot = './map_pngs/',
snapshot_name = 'Goldfish in British Columbia')

map_with_png

}
