% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmodel_r6.R
\name{pmodel}
\alias{pmodel}
\title{Title Set up a new priorization model.}
\value{
An R6 prioritization model model; use methods add(), run() and print() for more info.
}
\description{
Title Set up a new priorization model.

Title Set up a new priorization model.
}
\examples{
/dontrun{
No examples yet.
}
Title Print details of prioritization model


## ------------------------------------------------
## Method `pmodel$print`
## ------------------------------------------------

\dontrun{}
Title Add layer to prioritization model.


## ------------------------------------------------
## Method `pmodel$add`
## ------------------------------------------------

\dontrun{}
Title Run the prioritization model


## ------------------------------------------------
## Method `pmodel$run`
## ------------------------------------------------

\dontrun{}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pmodel_r6-print}{\code{pmodel$print()}}
\item \href{#method-pmodel_r6-add}{\code{pmodel$add()}}
\item \href{#method-pmodel_r6-run}{\code{pmodel$run()}}
\item \href{#method-pmodel_r6-clone}{\code{pmodel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pmodel_r6-print"></a>}}
\if{latex}{\out{\hypertarget{method-pmodel_r6-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmodel$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Details of prioritization model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{}
Title Add layer to prioritization model.

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pmodel_r6-add"></a>}}
\if{latex}{\out{\hypertarget{method-pmodel_r6-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmodel$add(role, x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{One of 'occurrence', 'geog_units', 'geog_id_col', 'risk', or 'weights'}

\item{\code{x}}{The layer to add (can be with or without geometry column)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Adds a layer to R6 prioritization model
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{}
Title Run the prioritization model

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pmodel_r6-run"></a>}}
\if{latex}{\out{\hypertarget{method-pmodel_r6-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmodel$run(n_bins = 3, quiet = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_bins}}{Number of bins to group risk layers and final risk estimates}

\item{\code{quiet}}{Should this function return copious feedback during the run?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prioritization model will have a new slot populated: geog_units_
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pmodel_r6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pmodel_r6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmodel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
