% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization_model.R
\name{prioritization_model}
\alias{prioritization_model}
\title{Title Run a prioritization model for a species of interest}
\usage{
prioritization_model(
  geog_units,
  geog_id_col = NULL,
  risk_factors = c(NA),
  risk_factor_weights = c(NA),
  n_bins = 3,
  data_folder = NULL,
  output_folder = NULL,
  set_wd_dir = NULL,
  species_name = NULL,
  plot_types = c("none"),
  save_spatial_table = FALSE,
  apply_second_bin_to_summed_bins = TRUE,
  ggplot_types = "bar_graph",
  quiet = F
)
}
\arguments{
\item{geog_units}{A spatial table (sf object) of polygons for which to estimate risk values}

\item{geog_id_col}{One or more columns that specify the unique identity of each polygon}

\item{risk_factors}{One or more variables pertaining to geographic units that could
influence the risk estimation; options include 'prox_to_settlements', 'rec_facilities'}

\item{risk_factor_weights}{A vector of weights for the selected risk factors; number of weights must be equal to number of risk factors}

\item{n_bins}{Number of bins to categorize risk factors into; defaults to 3}

\item{data_folder}{Folder on local machine for risk factor dataset(s)}

\item{output_folder}{Folder on local machine in which to put results}

\item{set_wd_dir}{set working directory for function; potentially temporary}

\item{species_name}{Species of interest's common name; will be intuited from data if not provided.}

\item{plot_types}{One or more plots to be generated, including 'static', 'leaflet', and 'plotly'}

\item{save_spatial_table}{TRUE or FALSE; save results table with geometries to output folder?}

\item{apply_second_bin_to_summed_bins}{TRUE or FALSE; bin the sum of risk factor bins? This gives us a final risk estimate column with values of 1, 2 or 3.}

\item{ggplot_types}{A vector of plot types to make ggplots for; currently only bar_graph is implemented}

\item{quiet}{Whether or not this function will return copious feedback at each step}
}
\value{
A table of estimated overall risk to ecosystem function based on the risk factors supplied; additional possible outputs include one or more choropleth images and the corresponding polygon dataset
}
\description{
Title Run a prioritization model for a species of interest
}
