% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_input_data.R
\name{process_input_data}
\alias{process_input_data}
\title{Title Read in and process additional spatial layers that represent either a risk increase
or a risk decrease to produce an output for left_joining with geog_units}
\usage{
process_input_data(
  input_name,
  input_weight,
  input_type,
  n_bins,
  geog_units,
  geog_id_col,
  data_folder,
  quiet
)
}
\arguments{
\item{input_name}{name of this potential risk factor for prioritization model}

\item{input_weight}{weight assigned to this potential risk factor (can be positive or negative)}

\item{n_bins}{number of bins into which to categorize this input}

\item{geog_units}{the sf object (i.e. a spatial table) for which we are
performing the prioritization model}

\item{geog_id_col}{the column of the sf object that identifies each row uniquely}

\item{data_folder}{folder on local machine wherein the risk factor
spatial files can be found}

\item{quiet}{should the function be quiet instead of chatty?}
}
\value{
internal function used in prioritization model to read in and process
additional risk factor layers.
}
\description{
Title Read in and process additional spatial layers that represent either a risk increase
or a risk decrease to produce an output for left_joining with geog_units
}
